Admin styling

+++ TwitterWall/src/TwitterWall/wwwroot/app/Components/mainadmin.component.ts
@@ -6,22 +6,64 @@ import { TwitterLogin } from "../Services/twitterlogin.service";
 @Component({
     selector: "main-admin",
     template: `
-        <h2>Main admin page</h2>
+        <div class="container main-admin">
+            <h1 class='text-center'>Admin</h1>
         {{errorMessage}}
-        <div *ngIf="!loggedIn">
-            <p>You need to login to view this page</p>
-            <button (click)="login()">Login</button>
-        </div>
-        <div *ngIf="loggedIn">
-            <p *ngIf="!loaded">Loading...</p>
-            <div *ngIf="loaded">
-                <ul>
-                    <li *ngFor="let event of events; let i=index">
-                        {{ event.Name }} (<a href="#/admin/{{event.Name}}">Admin</a>, <a href="#/events/{{event.Name}}">Visit</a>) <input type="button" value="Delete" (click)="removeEvent(i)">
-                    </li>
-                </ul>
-                <input type="text" placeholder="Event name" [(ngModel)]="inputEventName">
-                <input type="button" value="Create event" (click)="addEvent()">
+            <div *ngIf="!loggedIn">
+                <p class="lead">You need to login to view this page</p>
+                <div
+                    class="btn btn-primary btn-block btn-lg login"
+                    (click)="login()">
+                    <div *ngIf="!loggingIn">
+                        <img src="/img/Twitter_Social_Icon_White.png" />
+                        Login
+                    </div>
+                    <div *ngIf="loggingIn">
+                        <div class="spinner spinner-sm spinner-light"></div>
+                    </div>
+                </div>
+            </div>
+            <div *ngIf="loggedIn">
+                <div class="events-loading" *ngIf="!loaded">
+                    <div class="spinner"></div>
+                </div>
+                <div *ngIf="loaded">
+                    <ul class="list-group events">
+                        <li
+                            class="list-group-item event"
+                            *ngFor="let event of events; let i=index">
+                            <div class="event-name">{{ event.Name }}</div>
+                            <div class="event-actions">
+                                <a
+                                    class="text-info"
+                                    href="#/admin/{{event.Name}}">
+                                    Manage
+                                </a>
+                                <a
+                                    class="text-success"
+                                    href="#/events/{{event.Name}}">
+                                    View wall
+                                </a>
+                                <a
+                                    class="text-danger"
+                                    (click)="removeEvent(i)">
+                                    <span class="glyphicon glyphicon-trash" aria-hidden="true"></span>
+                                </a>
+                            </div>
+                        </li>
+                    </ul>
+                    <input
+                        class="form-control input-lg form-group"
+                        type="text"
+                        placeholder="Event name"
+                        [(ngModel)]="inputEventName"
+                    />
+                    <button
+                        class="btn btn-primary btn-block btn-lg"
+                        (click)="addEvent()">
+                        CreateEvent
+                    </button>
+                </div>
             </div>
         </div>
         `
@@ -31,6 +73,7 @@ export class MainAdminComponent {
     inputEventName: string;
     loaded: boolean = false;
     loggedIn: boolean = false;
+    loggingIn: boolean = false;
     errorMessage: string;
 
     constructor(private eventService: EventService, private twitterLogin: TwitterLogin) {
@@ -48,6 +91,7 @@ export class MainAdminComponent {
 
         if (window.sessionStorage.getItem("token") !== null && window.sessionStorage.getItem("handle") !== null) {
             this.loggedIn = true;
+            this.loggingIn = false;
         }
     }
 
@@ -70,10 +114,14 @@ export class MainAdminComponent {
     }
 
     removeEvent(index: number): void {
-        this.eventService.removeEvent(this.events[index].Id);
+        const result = confirm("Are you sure you want to remove this event?");
+        if (result) {
+          this.eventService.removeEvent(this.events[index].Id);
+        }
     }
 
     login(): void {
         this.twitterLogin.login();
+        this.loggingIn = true;
     }
 }
